refactor: move upscale test with other orc tests
